POLIO-2024: deduplicate stock tests
